requirements.txt
asgiref==3.3.4
async-timeout==4.0.2
Deprecated==1.2.13
Django==3.2.4
django-filter==2.4.0
django-redis==5.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
importlib-metadata==4.5.0
Markdown==3.3.4
packaging==21.3
PyJWT==1.7.1
pyparsing==3.0.9
pytz==2021.1
redis==4.3.4
sqlparse==0.4.1
typing-extensions==3.10.0.0
wrapt==1.14.1
zipp==3.4.1
